src/requirements.txt
# Frontend
streamlit==1.44.1
audio_recorder_streamlit
requests
speechrecognition==3.10.1
pyngrok
python-dotenv

# Backend
fastapi==0.115.5
uvicorn
python-multipart

# AI/ML
numpy==1.26.4
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
git+https://github.com/openai/whisper.git

# LangChain & Agents
langgraph==0.3.33
langchain-openai
langchain-huggingface
#vllm
#outlines==0.0.44
pydantic-settings

# Extra index for PyTorch GPU
--extra-index-url https://download.pytorch.org/whl/cu121
